+++ module2/gift-certificate-system/web/src/main/java/com/epam/esm/controller/UserController.java
@@ -49,10 +49,10 @@ public class UserController {
      */
     @ResponseStatus(HttpStatus.OK)
     @GetMapping(value = "/users", produces = {"application/hal+json"})
-    public CollectionModel<User> findAllUsers(@RequestParam(value = "page") Optional<Integer> page,
-                                              @RequestParam(value = "page_size") Optional<Integer> pageSize) {
-        int pageNumber = page.orElse(DEFAULT_PAGE_NUMBER);
-        int pageSizeNumber = pageSize.orElse(DEFAULT_PAGE_SIZE);
+    public CollectionModel<User> findAllUsers(@RequestParam(value = "page") Optional<Long> page,
+                                              @RequestParam(value = "page_size") Optional<Long> pageSize) {
+        Long pageNumber = page.orElse(DEFAULT_PAGE_NUMBER);
+        Long pageSizeNumber = pageSize.orElse(DEFAULT_PAGE_SIZE);
 
         ValidationUtils.checkPaginationData(pageNumber, pageSizeNumber);
